src/main/java/lv/javaguru/student_daniels_demishins/hw_lesson_5/lvl_5/ArrayUtil.java
package main.java.lv.javaguru.student_daniels_demishins.hw_lesson_5.lvl_5;

import java.util.Random;

public class ArrayUtil {

    public int[] createArray(int arrayLength) {
        return new int[arrayLength];
    }

    public  void fillArrayWithRandomNumbers(int[] array) {
        Random random = new Random();
        for (int i = 0; i < array.length; i++) {
            array[i] = random.nextInt(10);
        }
    }
    /* Task 34 */
    public void printArrayYoConsole(int[] array){
        Random random = new Random();
        for (int i = 0; i < array.length; i++) {
            array[i] = random.nextInt(10);
        }
    }

    public int findMaxNumber(int[] array){
        int max = array[0];
        for (int i = 0; i < array.length; i++) {
            if (array[i] > max) {
                max = array[i];
            }
        }
        return max;
    }



}
